refactor: modularize modal and list handling using composables
- Refactored App.vue to use useModal and useLists composables, simplifying modal state and list management.
- Updated ListCard.vue, ModalDialogue.vue, and TaskCard.vue to include comments explaining the purpose of each component.
- Fixed event emission in ModalDialogue.vue by passing the saved card through emit custom event: event.
- Optimized the card saving logic in useModal to manage card additions and updates more cleanly.
- Updated TaskCard.vue to re-arrange template for better readability and maintainability.
- Added type comments in types.ts to clarify the purpose of the application's types.
